WIP - Faster IT by starting SQ once + improvements here and there
