+++ its/src/test/kotlin/org/sonarlint/intellij/its/fixtures/IdeStatusBarFixture.kt
@@ -26,13 +26,14 @@ import com.intellij.remoterobot.fixtures.ComponentFixture
 import com.intellij.remoterobot.fixtures.DefaultXpath
 import com.intellij.remoterobot.fixtures.FixtureName
 import com.intellij.remoterobot.search.locators.byXpath
+import java.time.Duration
 
 @DefaultXpath(by = "IdeStatusBarImpl type", xpath = "//div[@class='IdeStatusBarImpl']")
 @FixtureName(name = "Status Bar")
 class IdeStatusBarFixture(remoteRobot: RemoteRobot, remoteComponent: RemoteComponent) : CommonContainerFixture(remoteRobot, remoteComponent) {
   val backgroundTaskPendingIcon
-    get() = find(ComponentFixture::class.java, byXpath("//div[@class='AsyncProcessIcon']"))
+    get() = find(ComponentFixture::class.java, byXpath("//div[@class='AsyncProcessIcon']"), Duration.ofSeconds(1))
 
   val pauseButton
-    get() = find(ComponentFixture::class.java, byXpath("//div[@accessiblename='Pause' and @class='InplaceButton']"))
+    get() = find(ComponentFixture::class.java, byXpath("//div[@accessiblename='Pause' and @class='InplaceButton']"), Duration.ofSeconds(1))
 }
